requirements.txt
Django==4.0.4
dj-rest-auth==2.2.5
django-allauth==0.51.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-filter==22.1
django-jazzmin==2.5.0
django-rest-auth==0.9.5
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gunicorn==20.1.0
Pillow==9.2.0
